+++ API/src/main/java/net/juligames/core/api/config/mapbacked/MapPart.java
@@ -1,8 +1,9 @@
-package net.juligames.core.api.config.env;
+package net.juligames.core.api.config.mapbacked;
 
 import org.jetbrains.annotations.ApiStatus;
 
 /**
+ * @apiNote should only be used with the {@link MapFeedInterpreter}
  * @author Ture Bentzin
  * 07.03.2023
  */
